feat: redesign as lovely-mermaid
Rename the package and repo (grok-mermaid -> lovely-mermaid) and move the
library into packages/lovely-mermaid as a bun workspace member; tools stay
at the repo root, a demo/ sibling comes later.

Split parse.ts into statements.ts + diagrams/{flowchart,state,class,er,
sequence}.ts + registry.ts; diagramKind and render dispatch derive from one
table. ClassInfo parallel arrays replaced by Node.sections.

Make every grammar lenient: unreadable statements drop with a warning and
the rest renders; bad block openers swallow their body; the one-final-line
salvage retry is gone. Node/edge caps truncate with a warning instead of
refusing (canvas cell cap stays fatal). Headers match exactly.

Rename Cls/span.cls to Role/span.role and surface author classes:
:::name and class statements land on span.classes, classDef declarations
parse into art.classDefs. toAnsi stays role-only.

191 tests; differential: 0 regressions (4274 identical / 2353 expected /
135 lenient / 8 slack / 410 trimmed).
