+++ biome.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://biomejs.dev/schemas/2.5.5/schema.json",
   "vcs": { "enabled": true, "clientKind": "git", "useIgnoreFile": true },
-  "files": { "includes": ["**", "!dist", "!src/width-data.ts"] },
+  "files": { "includes": ["**", "!**/dist", "!packages/lovely-mermaid/src/width-data.ts"] },
   "formatter": {
     "indentStyle": "space",
     "indentWidth": 2,
