+++ packages/lovely-mermaid/docs/streaming.gif
