+++ packages/lovely-mermaid/src/width-data.ts
