+++ packages/lovely-mermaid/src/width.ts
