+++ packages/lovely-mermaid/test/labels.test.ts
