+++ packages/lovely-mermaid/tsconfig.json
